pyproject.toml
[tool.poetry]
name = "sud"
version = "0.1.0"
description = "Scaleway Update DNS"
authors = ["Francesco Faraone"]
license = "Apache-2.0"
readme = "README.md"
packages = [
    { include = "sud" },
]


documentation = "https://github.com/ffaraone/sud"
homepage = "https://github.com/ffaraone/sud"
repository = "https://github.com/ffaraone/sud"

classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: Name Service (DNS)",
    "Topic :: System :: Networking",
    "Topic :: Utilities",
    "License :: OSI Approved :: Apache Software License"
]

keywords = ["sud", "scaleway", "dns", "dynamic", "dynamicdns"]



[tool.poetry.dependencies]
python = ">=3.10,<4"
requests = "^2.31.0"
click = "^8.1.7"
rich = "^13.7.0"
pyyaml = "^6.0.1"
python-telegram-bot = "^20.8"
humanize = "^4.9.0"

[tool.poetry.scripts]
sud = "sud.cli:main"

[tool.poetry.group.dev.dependencies]
ipython = "^8.21.0"
pytest = "^8.0.1"
black = "^24.2.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
flake8-pytest-style = "^1.7.2"
flake8-bugbear = "^24.2.6"
flake8-debugger = "^4.1.2"
pytest-randomly = "^3.15.0"
pytest-cov = "^4.1.0"
flake8-black = "^0.3.6"
flake8-isort = "^6.1.1"
pytest-mock = "^3.12.0"
responses = "^0.25.0"
pytest-deadfixtures = "^2.2.1"
mypy = "^1.8.0"
types-requests = "^2.31.0.20240218"
types-pyyaml = "^6.0.12.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
testpaths = "tests"
pythonpath = "."
addopts = "--cov=sud --cov-report=term-missing --cov-report=html --cov-report=xml"
log_cli = false

[tool.coverage.run]
branch = true

[tool.isort]
profile = "black"

[tool.black]
line-length = 79

[tool.mypy]
warn_no_return = false
